play the animation of clearing all displays when their isn't empty

+++ app/src/main/java/com/example/danilochagov/calc_3000/Main/MainActivity.java
@@ -298,7 +298,9 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void onDelAll(View view) {
-        deleteAllWithAnimation();
+        if (!main_display.getText().toString().equals("") || !old_display.getText().toString().equals("")) {
+            deleteAllWithAnimation();
+        }
     }
 
     private void deleteAllWithAnimation() {
